requirements.txt
pyserial
time
matplotlib.pyplot
scipy.ndimage
